~ | Removed app; only transcendence
