+++ transcendence/models.py
@@ -1,5 +1,5 @@
-from django.db import models
 from django.conf import settings
+from django.db import models
 
 # Extends django.contrib.auth User
 class Profile ( models.Model ):
