+++ transcendence/templates/app/components/_footer.html
