+++ transcendence/templates/app/pages/logout.html
