+++ transcendence/templates/app/pages/play.html
