+++ transcendence/templates/app/pages/signup.html
