+++ transcendence/templates/app/pages/solo_play.html
