+++ transcendence/templates/app/pages/tournament.html
