+++ transcendence/wsgi.py
@@ -1,6 +1,5 @@
-# WSGI config for transcendence project.
-
 import os
+
 from django.core.wsgi import get_wsgi_application
 
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'transcendence.settings')
